requirements.txt
yarl<1.2
discord.py~=1.7.3
pandas~=0.23.0
beautifulsoup4~=4.6.0
pylint~=1.7.3
aiohttp>=3.7.4
opencensus-ext-azure
emoji
